[BE-4] move models to Eloquent folder
#4
