+++ app/Http/Controllers/InterestController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Interest;
+use App\Eloquent\Interest;
 use Illuminate\Http\Request;
 
 class InterestController extends Controller
@@ -41,7 +41,7 @@ class InterestController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Interest  $interest
+     * @param  \App\Eloquent\Interest  $interest
      * @return \Illuminate\Http\Response
      */
     public function show(Interest $interest)
@@ -52,7 +52,7 @@ class InterestController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Interest  $interest
+     * @param  \App\Eloquent\Interest  $interest
      * @return \Illuminate\Http\Response
      */
     public function edit(Interest $interest)
@@ -64,7 +64,7 @@ class InterestController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Interest  $interest
+     * @param  \App\Eloquent\Interest  $interest
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Interest $interest)
@@ -75,7 +75,7 @@ class InterestController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Interest  $interest
+     * @param  \App\Eloquent\Interest  $interest
      * @return \Illuminate\Http\Response
      */
     public function destroy(Interest $interest)
