+++ database/factories/UserFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
-use App\User;
+use App\Eloquent\User;
 use Illuminate\Support\Str;
 use Faker\Generator as Faker;
